+++ execution/requests/eip7002/withdrawal.go
@@ -18,25 +18,30 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package eip7685
+package eip7002
 
 import (
 	"bytes"
 	"context"
+	"encoding/binary"
 	"math/big"
 
 	"github.com/berachain/beacon-kit/errors"
 	beaconbytes "github.com/berachain/beacon-kit/primitives/bytes"
 	"github.com/berachain/beacon-kit/primitives/crypto"
+	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/ethereum/go-ethereum/params"
 )
 
-// GetConsolidationFee returns the consolidation fee in wei. See https://eips.ethereum.org/EIPS/eip-7251 for more.
-// This is only expected to be used in tests to form test scenarios.
-func GetConsolidationFee(ctx context.Context, client rpcClient) (*big.Int, error) {
+type feeOpts struct {
+	To string `json:"to"`
+}
+
+// GetWithdrawalFee returns the withdrawal fee in wei. See https://eips.ethereum.org/EIPS/eip-7002 for more.
+func GetWithdrawalFee(ctx context.Context, client rpcClient) (*big.Int, error) {
 	var result string
 	feeInput := &feeOpts{
-		To: params.ConsolidationQueueAddress.String(),
+		To: params.WithdrawalQueueAddress.String(),
 	}
 	err := client.Call(ctx, &result, "eth_call", feeInput)
 	if err != nil {
@@ -49,15 +54,15 @@ func GetConsolidationFee(ctx context.Context, client rpcClient) (*big.Int, error
 	return n, nil
 }
 
-// CreateConsolidationRequestData returns the request body formatted as defined by the EIP-7251 specification.
-// This is only expected to be used in tests to form test scenarios.
-func CreateConsolidationRequestData(sourcePubKey, targetPubKey crypto.BLSPubkey) (beaconbytes.Bytes, error) {
+// CreateWithdrawalRequestData returns the request body formatted as defined by the EIP-7002 specification.
+func CreateWithdrawalRequestData(blsPubKey crypto.BLSPubkey, withdrawAmount math.Gwei) (beaconbytes.Bytes, error) {
 	// Create a buffer to hold the packed encoding.
 	var packed bytes.Buffer
-	if _, err := packed.Write(sourcePubKey[:]); err != nil {
+	if _, err := packed.Write(blsPubKey[:]); err != nil {
 		return nil, err
 	}
-	if _, err := packed.Write(targetPubKey[:]); err != nil {
+	// Write the uint64 value in big-endian order.
+	if err := binary.Write(&packed, binary.BigEndian, withdrawAmount); err != nil {
 		return nil, err
 	}
 	return packed.Bytes(), nil
